+++ examples/re-lb-waap-bot-cdn/4-cdn-lb.tf
@@ -5,7 +5,7 @@ resource "volterra_cdn_loadbalancer" "cdn-distri-app1-tf" {
   name      = "cdn-distri-app1-tf"
   namespace = var.f5xc_namespace
 
-  domains = ["apptf1-cdn.wwt.xcsdemo.com"]
+  domains = ["cdn-${var.app_domain}"]
   
   https_auto_cert {
     add_hsts = true
@@ -17,18 +17,26 @@ resource "volterra_cdn_loadbalancer" "cdn-distri-app1-tf" {
     origin_request_timeout = "100s"
 
     public_name {
-        dns_name = "cdnapp.hvf5lab.com"
+        dns_name = var.app_domain
         refresh_interval = "20"
     }
 
     origin_servers {
         public_name {
-           dns_name = "cdnapp.hvf5lab.com"
+           dns_name = var.app_domain
            refresh_interval = "20"
         }
       }
     // One of the arguments from this list "no_tls use_tls" must be set
-    no_tls = true
+    # no_tls = true
+    use_tls {
+      use_host_header_as_sni = true
+      tls_config {
+          default_security = true
+      }
+      skip_server_verification = true
+      no_mtls = true
+    } 
   }
   custom_cache_rule {
     cdn_cache_rules {
@@ -50,4 +58,10 @@ resource "volterra_cdn_loadbalancer" "cdn-distri-app1-tf" {
 
   }
 
+  active_service_policies {
+    policies {
+      name      = volterra_service_policy.cdn-sp.name
+      namespace = var.f5xc_namespace
+    }
+  }
 }