+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a detailed "Bot Protection Clarification" section to the main `README.md` and relevant example documentation.
 
 ### Changed
+- Updated the `re-lb-waap-bot-cdn` example to include a service policy and updated the routes.
 - Updated each example documention with the provider requirements to show the verion used to test the exmaple
 - Updated all examples to use variables for `app_domain`, `site_name`, and `origin_dns_name`.
 - Updated the `ce-lb-waap-bot` example to use a custom certificate.